Übung 4 Leonard Richertz
